consolidate ctf util code with class into a single file
