+++ src/aspire/ctf/__init__.py
@@ -1,2 +1,1 @@
-from .ctf_estimator import CtfEstimator
-from .estimate_ctf import estimate_ctf
+from .ctf_estimator import CtfEstimator, estimate_ctf
